requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chumpy==0.70
click==8.2.1
cmake==3.25.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
crcmod==1.7
cryptography==45.0.3
cycler==0.12.1
Cython==3.1.1
dashscope==1.19.3
dashscope==1.19.3
debugpy==1.8.14
decorator==5.2.1
dnspython==2.7.0
docker-pycreds==0.4.0
einops==0.7.0
eventlet==0.40.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
filelock==3.14.0
filterpy==1.4.5
Flask==3.1.1
Flask-SocketIO==5.5.1
fonttools==4.48.1
fsspec==2024.2.0
gitdb==4.0.12
GitPython==3.1.44
gmpy2==2.2.1
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.2
httptools==0.6.4
huggingface-hub==0.30.0
idna==3.6
imageio==2.33.1
importlib-resources==6.1.1
intel-cmplr-lib-ur==2024.2.1
intel-openmp==2024.2.1
ipykernel==6.29.5
ipython==8.36.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.1
json-tricks==3.17.3
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.5
lazy_loader==0.3
lit==15.0.7
llvmlite==0.42.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2024.2.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
# mmcv==2.2.0
# mmdet==3.1.0
# mmengine==0.10.7
# mmpose==1.3.2
model-index==0.1.11
mpmath==1.3.0
munkres==1.1.4
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.1.4
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
pip-install==1.3.5
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycocotools==2.0.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.1.0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2023.4
PyYAML==6.0.1
pyzmq==26.4.0
regex==2024.11.6
requests==2.28.2
rich==13.4.2
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
sentry-sdk==2.29.1
setproctitle==1.3.6
shapely==2.1.1
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
tabulate==0.9.0
tbb==2021.13.1
termcolor==3.1.0
terminaltables==3.1.10
threadpoolctl==3.6.0
tifffile==2024.1.30
timm==0.9.12
tokenizers==0.21.1
tomli==2.2.1
torch==2.5.1
torch-dct==0.1.6
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.5.1
tqdm==4.65.2
traitlets==5.14.3
transformers==4.52.3
triton==3.1.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.3
uvloop==0.21.0
wandb==0.19.11
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xtcocotools==1.14.3
yacs==0.1.8
yapf==0.43.0
zipp==3.17.0
flask_sqlalchemy
flask_wtf
mysql-connector-python
PyMySQL
gunicorn
flask_compress
openai
faiss-cpu